backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow-cpu==2.16.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
reportlab==4.0.9
openai==1.12.0
h5py==3.10.0